fixed information loss when pausing tasks
Pausing tasks stored a rational of DiffTime to represent the amount of
time that was accumulated on an unfinished task by it being paused; this
behavior is wrong due to it not being able to tell you when that time
occured. To fix this we still have a representation of DiffTime through
TimeDiff, but instead use TimeTo to represent at what time it was both
started and paused.